Create bangs.json (#11)
* Create bangs.json

Simple regex to create "bang" for different services. A concept introduced by DDG (https://duckduckgo.com/bangs)
It can be very useful for example if you get a text you about a movie just share the name to URLCheck and add !i for IMDb Search
It is simple and offline.